% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transport-simplex.R
\name{counterfactual_w}
\alias{counterfactual_w}
\title{Builds counterfactuals of observations from group 0 to group 1 using an
estimated mapping.}
\usage{
counterfactual_w(mapping, X0, X1)
}
\arguments{
\item{mapping}{Wasserstein mapping between the two sets of probability
vectors}

\item{X0}{Matrix (or data frame) with vectors of probabilities in group 0.}

\item{X1}{Matrix (or data frame) with vectors of probabilities in group 1.}
}
\value{
A matrix with the counterfactual values using the mapping.
}
\description{
Builds counterfactuals of observations from group 0 to group 1 using an
estimated mapping.
}
\examples{
data(toydataset)
# Observations in group 0
X0 <- toydataset[toydataset$group == 0, c("A", "B", "C")]
# Observations in group 1
X1 <- toydataset[toydataset$group == 1, c("A", "B", "C")]

# Optimal Transport using Linear Programming:
mapping <- wasserstein_simplex(as.matrix(X0), as.matrix(X1))

# The counterfactuals of observations of group 0 in group 1
counterfactuals_0_1 <- counterfactual_w(mapping, X0, X1)
head(counterfactuals_0_1)
}
