% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transport-simplex.R
\name{get_ot_mapping}
\alias{get_ot_mapping}
\title{Learn OT mapping from group 0 to group 1 based on the representation of the
composition data from the simplex (\eqn{S_d}) to a \eqn{d-1} Euclidean space.}
\usage{
get_ot_mapping(X0, X1, isomorphism = c("clr", "alr", "ilr"))
}
\arguments{
\item{X0}{Data frame with observations in group 0.}

\item{X1}{Data frame with observations in group 1.}

\item{isomorphism}{Isomorphism used to map the composition data from the
simplex (\eqn{S_d}) to a \eqn{d-1} Euclidean space. Three possibilities: \code{"alr"}
(additive log ratio transform), \code{"clr"} (centered log ratio transform),
and \code{"ilr"} (isometric log ratio transform).

@returns A list with the following elements:
\itemize{
\item \code{m0}, \code{m1}: empirical mean of \eqn{z_0} and \eqn{z_1} (vectors of orthonormal
coordinates).
\item \code{S0}, \code{S1}: empirical covariance matrices of \eqn{z_0} and \eqn{z_1}.
\item \code{A}: symmetric positive matrix that satisfies
\eqn{\bf{A}\bf{\Sigma}_{0}\bf{A}=\bf{\Sigma}_{1}}
\item \code{isomorphism}: name of the isomorphism used (alr, clr, irl).
}}
}
\description{
Learn OT mapping from group 0 to group 1 based on the representation of the
composition data from the simplex (\eqn{S_d}) to a \eqn{d-1} Euclidean space.
}
\examples{
# First three columns: probabilities of being of class A, B, or C.
# Last column: group (0 or 1)
data(toydataset)
X0 <- toydataset[toydataset$group == 0, c("A", "B", "C")]
X1 <- toydataset[toydataset$group == 1, c("A", "B", "C")]
ot_mapping <- get_ot_mapping(X0 = X0, X1 = X1, isomorphism = "clr")
ot_mapping$A
}
