% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transport-simplex.R
\name{interpolated}
\alias{interpolated}
\title{Extract the interpolated values of transported vectors of compositional data}
\usage{
interpolated(x)
}
\arguments{
\item{x}{Transported compositional data.}
}
\value{
A list with the interpolated values. Each element contains a tibble
giving the interpolated values for an observation.
}
\description{
Extract the interpolated values of transported vectors of compositional data
}
\seealso{
\code{\link[=transport_simplex]{transport_simplex()}}
}
