% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toydataset}
\alias{toydataset}
\title{Toydataset}
\format{
\subsection{\code{toydataset}}{

A data frame with 97 rows and 4 columns:
\describe{
\item{A}{Probability of having characteristic A}
\item{B}{Probability of having characteristic B}
\item{C}{Probability of having characteristic C}
\item{group}{Group: either 0 or A}
}
}
}
\usage{
toydataset
}
\description{
A toy dataset with individuals from two groups: 0 and 1. The
individuals have a fictive attribute with three levels: A, B, or C. The
table contains the probability for each levels of the attribute.
}
\keyword{datasets}
