% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transport-simplex.R
\name{transport_simplex_new}
\alias{transport_simplex_new}
\title{Transport new compositional data from group 0 to group 1 using a previously
learned mapping.}
\usage{
transport_simplex_new(transport, newdata, n_interp = 1)
}
\arguments{
\item{transport}{Previously learned mapping.}

\item{newdata}{New data frame with composition data.}

\item{n_interp}{Number of steps in the interpolation (default to 1: no
interpolation).}
}
\value{
A tibble with the transported values. If \code{n_interp} is larger than
1, the result also contains a list with the interpolated values, in the
\code{"interpolated"} attribute ; else, the attribute is \code{NULL}. The attribute
\code{"ot_mapping"} stores the mapping.
}
\description{
Transport new compositional data from group 0 to group 1 using a previously
learned mapping.
}
\examples{
# First three columns: probabilities of being of class A, B, or C.
# Last column: group (0 or 1)
data(toydataset)
X0 <- toydataset[toydataset$group == 0, c("A", "B", "C")]
X1 <- toydataset[toydataset$group == 1, c("A", "B", "C")]

# Transport only, from group 0 to group 1, using centered log ratio transform:
transp <- transport_simplex(X0 = X0, X1 = X1, isomorphism = "clr")
head(transp)

# If we want to transport new points:
new_obs <- data.frame(A = c(.2, .1), B = c(.6, .5), C = c(.2, .4))
transp_new_obs <- transport_simplex_new(transport = transp, newdata = new_obs)
transp_new_obs

# If we want to get interpolated values using McCann (1997) displacement
# interpolation: (here, with 5 intermediate points)
transp_new_obs_with_interp <- transport_simplex_new(
  transport = transp, newdata = new_obs, n_interp = 5
)
interpolated(transp_new_obs_with_interp)[[1]] # first new obs
interpolated(transp_new_obs_with_interp)[[2]] # second new obs
}
\references{
McCann, Robert J. 1997. "A Convexity Principle for Interacting
Gases." Advances in Mathematics 128 (1): 153–79.
}
\seealso{
\code{\link[=transport_simplex]{transport_simplex()}}
}
