% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transport-simplex.R
\name{transport_x_alr}
\alias{transport_x_alr}
\title{Gaussian-based transport of a single observation from group 0 to group 1
on \eqn{S_d}, using additive log ratio transform}
\usage{
transport_x_alr(x, n_interp = 31, A, m0, m1)
}
\arguments{
\item{x}{Value to transport.}

\item{n_interp}{Number of points for the interpolation.}

\item{A}{Symmetric positive matrix from OT.}

\item{m0}{Average of transformed values in group 0.}

\item{m1}{Average of transformed values in group 1.}
}
\description{
Gaussian-based transport of a single observation from group 0 to group 1
on \eqn{S_d}, using additive log ratio transform
}
