% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transport-simplex.R
\name{wasserstein_simplex}
\alias{wasserstein_simplex}
\title{Wasserstein distance between two sets of probability vectors X and Y}
\usage{
wasserstein_simplex(X, Y, wx = NULL, wy = NULL)
}
\arguments{
\item{X}{Matrix of probability vectors in a first group.}

\item{Y}{Matrix of probability vectors in a second group.}

\item{wx}{Weights (marginal distribution) for X. Default to \code{NULL} (uniform
weights will be used).}

\item{wy}{Weights (marginal distribution) for Y. Default to \code{NULL} (uniform
weights will be used).}
}
\value{
A list with two elements:
\itemize{
\item \code{distance}: the Wassterstein distance
\item \code{plan}: the optimal transport plan describing how mass is transported
between X and Y.
}
}
\description{
Wasserstein distance between two sets of probability vectors X and Y
}
